+++ config/routes.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+LlmCostTracker::Engine.routes.draw do
+  root "dashboard#index"
+end